requirements.txt
carla
pygame
numpy
Pillow